+++ crates/overlay/src/renderer/dx12.rs
@@ -292,7 +292,7 @@ impl Dx12Renderer {
         backbuffer_index: u32,
         render_target: D3D12_CPU_DESCRIPTOR_HANDLE,
         queue: &ID3D12CommandQueue,
-        position: (f32, f32),
+        position: (i32, i32),
         size: (u32, u32),
         screen: (u32, u32),
     ) -> anyhow::Result<()> {
@@ -343,8 +343,8 @@ impl Dx12Renderer {
         };
 
         let rect: [f32; 4] = [
-            (position.0 / screen.0 as f32) * 2.0 - 1.0,
-            -(position.1 / screen.1 as f32) * 2.0 + 1.0,
+            (position.0 as f32 / screen.0 as f32) * 2.0 - 1.0,
+            -(position.1 as f32 / screen.1 as f32) * 2.0 + 1.0,
             (size.0 as f32 / screen.0 as f32) * 2.0,
             -(size.1 as f32 / screen.1 as f32) * 2.0,
         ];
