+++ crates/overlay/src/renderer/opengl.rs
@@ -147,7 +147,7 @@ impl OpenglRenderer {
     #[tracing::instrument(skip(self))]
     pub fn draw(
         &mut self,
-        position: (f32, f32),
+        position: (i32, i32),
         size: (u32, u32),
         screen: (u32, u32),
     ) -> anyhow::Result<()> {
@@ -160,8 +160,8 @@ impl OpenglRenderer {
         };
 
         let rect: [f32; 4] = [
-            (position.0 / screen.0 as f32) * 2.0 - 1.0,
-            -(position.1 / screen.1 as f32) * 2.0 + 1.0,
+            (position.0 as f32 / screen.0 as f32) * 2.0 - 1.0,
+            -(position.1 as f32 / screen.1 as f32) * 2.0 + 1.0,
             (size.0 as f32 / screen.0 as f32) * 2.0,
             -(size.1 as f32 / screen.1 as f32) * 2.0,
         ];
